+++ src/main/java/nextstep/ladder/LadderController.java
@@ -1,5 +1,8 @@
 package nextstep.ladder;
 
+import nextstep.ladder.domain.player.Players;
+import nextstep.ladder.ui.InputView;
+
 public class LadderController {
 
     public static void main(String[] args) {
