+++ src/main/java/nextstep/ladder/domain/player/Player.java
@@ -1,4 +1,4 @@
-package nextstep.ladder;
+package nextstep.ladder.domain.player;
 
 public class Player {
     private final String name;
