+++ src/main/java/nextstep/ladder/domain/player/Players.java
@@ -1,4 +1,4 @@
-package nextstep.ladder;
+package nextstep.ladder.domain.player;
 
 import java.util.List;
