+++ src/test/java/nextstep/ladder/domain/PlayersTest.java
@@ -1,5 +1,6 @@
-package nextstep.ladder;
+package nextstep.ladder.domain;
 
+import nextstep.ladder.domain.player.Players;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
